11-2 - рефакторинг
