--- src/main/kotlin/lesson_11/User2.kt
@@ -1,13 +1,0 @@
-package lesson_11
-
-class User2(
-    val uid: String,
-    val login: String,
-    var password: String,
-    val email: String,
-    var bio: String = "",
-) {
-    override fun toString(): String {
-        return "User(uid='$uid', login='$login', password='$password', email='$email', bio='$bio')"
-    }
-}
