+++ src/main/kotlin/lesson_11/lesson11_task2.kt
@@ -1,5 +1,18 @@
 package lesson_11
 
+class User2(
+    val uid: String,
+    val login: String,
+    var password: String,
+    val email: String,
+    var bio: String = "",
+) {
+    override fun toString(): String {
+        return "User(uid='$uid', login='$login', password='$password', email='$email', bio='$bio')"
+    }
+}
+
+
 fun main() {
     val user1 = User2(
         uid = "1",
